Remove extension of active_filters to response.
Active filters are now destroyed as soon as the Response is received,
even if the body is still incoming. Prolonging their lifetime was a
workaround and is no longer necessary. Tasks related to filters such as
ExternalProcessor, manage their own lifetime by ignoring the closure of
the main channel until the body has been fully streamed in both
directions.